+++ src/StorageKey/UiAuditTrailStorageKey.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey;
+
+use Heptacom\HeptaConnect\Storage\Base\Contract\UiAuditTrailKeyInterface;
+
+final class UiAuditTrailStorageKey extends AbstractStorageKey implements UiAuditTrailKeyInterface
+{
+}